added Flag support to Jun compiler, changed Globals/Flags/Locals to use 16-bit offsets (must recompile all JC for this)

+++ DosjunEditor/Jun/Variable.cs
@@ -2,7 +2,7 @@
 {
     public class Variable
     {
-        public byte Index { get; set; }
+        public short Index { get; set; }
         public string Name { get; set; }
         public Scope Scope { get; set; }
